Move walkDirectory to pure files file
